requirements.txt
pandas
scikit-learn
flask
flask_cors
psycopg2-binary==2.9.5